Simialrly move block storage into separate module
