+++ hschain-PoW/HSChain/PoW/Consensus.hs
@@ -51,7 +51,6 @@ import Control.Monad
 import Control.Monad.Except
 import Control.Monad.State.Strict
 import Data.List          (sortOn)
-import Data.Foldable
 import Data.Typeable      (Typeable)
 import Data.Maybe
 import Data.Sequence      (Seq(Empty,(:<|),(:|>)),(|>))
@@ -67,6 +66,8 @@ import GHC.Generics (Generic)
 
 import HSChain.PoW.Types
 import HSChain.PoW.BlockIndex
+import HSChain.PoW.Store
+
 
 ----------------------------------------------------------------
 -- Block index
@@ -95,43 +96,6 @@ makeLocator  = Locator . takeH 10 . Just
 ----------------------------------------------------------------
 -- Blockchain state handling
 ----------------------------------------------------------------
-
--- | API for append only block storage. It should always contain
---   genesis block.
-data BlockDB m b = BlockDB
-  { storeBlock :: Block b -> m ()
-    -- ^ Put block into storage. It should be idempotent. That is
-    --   storing block already in storage should be a noop.
-  , retrieveBlock :: BlockID b -> m (Maybe (Block  b))
-    -- ^ Retrive complete block by its identifier
-  , retrieveHeader :: BlockID b -> m (Maybe (Header b))
-    -- ^ Retrieve header by its identifier
-  , retrieveAllHeaders :: m [Header b]
-    -- ^ Retrieve all headers from storage in topologically sorted
-    --   order. (Ordering block by height would achieve that for
-    --   example).
-  }
-
--- | Build block index from blocks
-buildBlockIndex :: (Monad m, BlockData b) => BlockDB m b -> m (BlockIndex b)
-buildBlockIndex BlockDB{..} = do
-  -- FIXME: decide what to do with orphan blocks
-  (genesis,headers) <- retrieveAllHeaders >>= \case
-    genesis:headers -> return (genesis,headers)
-    []              -> error "buildBlockIndex: no blocks in storage"
-  --
-  let idx0      = blockIndexFromGenesis genesis
-      add idx b@GBlock{..} = case (`lookupIdx` idx) =<< prevBlock of
-        Nothing     -> error "blockIndexFromGenesis: orphan block"
-        Just parent -> insertIdx BH
-          { bhHeight   = blockHeight
-          , bhTime     = blockTime
-          , bhBID      = blockID b
-          , bhWork     = bhWork parent <> blockWork b
-          , bhPrevious = Just parent
-          , bhData     = blockData
-          } idx
-  return $! foldl' add idx0 headers
 
 
 -- | View on state of blockchain. It's expected that store is backed
